Review_v4

+++ MyStore/Services/Categories/CategoryService.cs
@@ -3,6 +3,7 @@ using MyStore.Constant;
 using MyStore.DTO;
 using MyStore.Models;
 using MyStore.Repository.CategoryRepository;
+using MyStore.Storage;
 
 namespace MyStore.Services.Categories
 {
@@ -10,26 +11,52 @@ namespace MyStore.Services.Categories
     {
         private readonly ICategoryRepository _categoryRepository;
         private readonly IMapper _mapper;
+        private readonly IFileStorage _fileStorage;
+        private readonly string path = "assets/images/categories";
 
-        public CategoryService(ICategoryRepository categoryRepository, IMapper mapper)
+        public CategoryService(ICategoryRepository categoryRepository, IMapper mapper, IFileStorage fileStorage)
         {
             _categoryRepository = categoryRepository;
             _mapper = mapper;
+            _fileStorage = fileStorage;
         }
-        public async Task<CategoryDTO> AddCategoryAsync(string name)
+        public async Task<CategoryDTO> AddCategoryAsync(string name, IFormFile image)
         {
-            var category = new Category
+           try
             {
-                Name = name
-            };
-            await _categoryRepository.AddAsync(category);
-            return _mapper.Map<CategoryDTO>(category);
+                string fileName = Guid.NewGuid().ToString() + Path.GetExtension(image.FileName);
+                Category category = new()
+                {
+                    Name = name,
+                    ImageUrl = Path.Combine(path, fileName),
+                };
+                await _categoryRepository.AddAsync(category);
+                await _fileStorage.SaveAsync(path, image, fileName);
+                return _mapper.Map<CategoryDTO>(category);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.InnerException?.Message ?? ex.Message);
+            }
 
         }
 
         public async Task DeleteCategoryAsync(int id)
         {
-            await _categoryRepository.DeleteAsync(id);
+            try
+            {
+                var category = await _categoryRepository.FindAsync(id);
+                if (category != null)
+                {
+                    _fileStorage.Delete(category.ImageUrl);
+                    await _categoryRepository.DeleteAsync(id);
+                }
+                else throw new Exception($"ID {id} " + ErrorMessage.NOT_FOUND);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.InnerException?.Message ?? ex.Message);
+            }
         }
 
         public async Task<IEnumerable<CategoryDTO>> GetAllCategoriesAsync()
@@ -38,19 +65,27 @@ namespace MyStore.Services.Categories
             return _mapper.Map<IEnumerable<CategoryDTO>>(categories);
         }
 
-        public async Task<CategoryDTO> UpdateCategoryAsync(int id, string name)
+        public async Task<CategoryDTO> UpdateCategoryAsync(int id, string name, IFormFile? image)
         {
             var category = await _categoryRepository.FindAsync(id);
-            if (category == null)
-            {
-                throw new ArgumentException($"ID {id}" + ErrorMessage.NOT_FOUND);
-            }
-            else
+            if (category != null)
             {
                 category.Name = name;
+
+                if (image != null)
+                {
+                    _fileStorage.Delete(category.ImageUrl);
+
+                    string fileName = Guid.NewGuid() + Path.GetExtension(image.FileName);
+                    category.ImageUrl = Path.Combine(path, fileName);
+
+                    await _fileStorage.SaveAsync(path, image, fileName);
+                }
+                
                 await _categoryRepository.UpdateAsync(category);
                 return _mapper.Map<CategoryDTO>(category);
             }
+            else throw new ArgumentException($"ID {id}" + ErrorMessage.NOT_FOUND);
         }
     }
 }
